added option to disable state model module or to run the module in storeless mode
