+++ testar/src/nl/ou/testar/StateModel/StateModelManagerFactory.java
@@ -4,7 +4,6 @@ import es.upv.staq.testar.CodingManager;
 import nl.ou.testar.StateModel.ActionSelection.ActionSelector;
 import nl.ou.testar.StateModel.ActionSelection.CompoundFactory;
 import nl.ou.testar.StateModel.Event.StateModelEventListener;
-import nl.ou.testar.StateModel.Persistence.OrientDB.OrientDBManager;
 import nl.ou.testar.StateModel.Persistence.PersistenceManager;
 import nl.ou.testar.StateModel.Persistence.PersistenceManagerFactory;
 import nl.ou.testar.StateModel.Persistence.PersistenceManagerFactoryBuilder;
@@ -14,12 +13,16 @@ import org.fruit.monkey.ConfigTags;
 import org.fruit.monkey.Settings;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 public class StateModelManagerFactory {
 
     public static StateModelManager getStateModelManager(Settings settings) {
+        // first check if the state model module is enabled
+        if(!settings.get(ConfigTags.StateModelEnabled)) {
+            return new DummyModelManager();
+        }
+
         // check the attributes for the abstract state id
         if (settings.get(ConfigTags.AbstractStateAttributes).isEmpty()) {
             throw new RuntimeException("No Abstract State Attributes were provided in the settings file");
@@ -42,7 +45,7 @@ public class StateModelManagerFactory {
 
         // get a persistence manager
         PersistenceManagerFactoryBuilder.ManagerType managerType;
-        if (!settings.get(ConfigTags.StateModelEnabled)) {
+        if (settings.get(ConfigTags.DataStoreMode).equals(PersistenceManager.DATA_STORE_MODE_NONE)) {
             managerType = PersistenceManagerFactoryBuilder.ManagerType.DUMMY;
         }
         else {
@@ -68,7 +71,7 @@ public class StateModelManagerFactory {
                 persistenceManager instanceof StateModelEventListener ? (StateModelEventListener) persistenceManager : null);
         ActionSelector actionSelector = CompoundFactory.getCompoundActionSelector();
 
-        return new StateModelManager(abstractStateModel, actionSelector, persistenceManager, concreteStateTags, sequenceManager);
+        return new ModelManager(abstractStateModel, actionSelector, persistenceManager, concreteStateTags, sequenceManager);
     }
 
 }
